feature/event-handlers added + last review requested changes + eslint fixes
