+++ src/App.test.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { render } from '@testing-library/react';
-import App from './App';
+import App from './Components/App';
 
 test('renders learn react link', () => {
   const { getByText } = render(<App />);
