+++ src/style.css
@@ -4,22 +4,25 @@ body {
     -moz-osx-font-smoothing: grayscale;
     padding: 0;
     margin: 0;
-
+    display: flex;
+    justify-content: center;
+    background-image: radial-gradient(rgba(0,0,0,.94), rgba(0,0,0,.96));
+    background-repeat: no-repeat;
+    min-height: -webkit-fill-available;
 }
 
 div#root {
     display: flex;
     justify-content: center;
     padding-top: 5em;
-    position: fixed;
-    width: 100%;
-    height: 100%;
-    background-image: radial-gradient(rgba(0,0,0,.94), rgba(0,0,0,.96));
+
+    width: 700px;
 }
 
+
 .btn-0 {
-    min-width: 5em;
-    min-height: 5em;
+    min-width: 100px;
+    min-height: 100px;
     font-size: 1.375em;
     color: rgba(0,0,0,.74);
     background-image: none;
@@ -32,10 +35,10 @@ div#root {
 }
 
 .btn-0.size-2 {
-    min-width: 10em;
-    min-height: 5em;
+    min-width: 200px;
+    min-height: 100px;
 }
-.btn-0.last {
+.btn-0.orange {
     background-color: orange;
 }
 .display {
@@ -66,4 +69,5 @@ button:focus, input[type="button"]:active {
 
 .buttonpanel {
     border-left: 1px solid rgba(0,0,0,.15);
+    display: flexbox;
 }
